+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v.015
+- (New) Metrics class to keep track of different metrics while any algorithm is executing. Calculate hypervolume of any metric in given range.
 - (New) ArrayUtil::max, ArrayUtil::min util methods for int, long and double primitives
 - (New) Add option to choose how frequently any solution serializer should execute: best_per_instance, best_per_alg_instance and all iterations.
 - (Change) Improve SolutionSerializer::export method